+++ src/irmin-pack/inode.ml
@@ -50,6 +50,8 @@ module type S = sig
   val sync : ?on_generation_change:(unit -> unit) -> 'a t -> unit
 
   val clear_caches : 'a t -> unit
+
+  val copy_entry : offset:int64 -> length:int -> key -> 'a t -> 'a t -> unit
 end
 
 module type CONFIG = sig
@@ -824,5 +826,9 @@ struct
 
   let sync = Inode.sync
 
+  let clear = Inode.clear
+
   let clear_caches = Inode.clear_caches
+
+  let copy_entry = Inode.copy_entry
 end